% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagWhitelistGeneration.R
\name{AddCellTagFreqSort}
\alias{AddCellTagFreqSort}
\title{CellTag Frequency Sort Table}
\usage{
AddCellTagFreqSort(celltag.obj)
}
\arguments{
\item{celltag.obj}{A CellTag Object with CellTags extracted}
}
\value{
A CellTag Object with attribute (celltag.freq.stats) filled.
}
\description{
This function counts and sorts the identified CellTags from Fastq file
}
\examples{
CellTagWhitelistFiltering(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
