% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagNetworkContruction.R
\name{addData2Nodes}
\alias{addData2Nodes}
\title{Add Additional Information to the Nodes}
\usage{
addData2Nodes(celltag.obj, additional_data)
}
\arguments{
\item{celltag.obj}{A CellTag object with nodes filled}

\item{additional_data}{A data frame with auxillary information about the nodes (rownames = the nodes names)}
}
\value{
A CellTag object with the attribute (nodes) modified.
}
\description{
This function add auxillary information to the nodes. Such information can include cluster information, cell type information and so on. The information should be stored as a data frame when passing in to the funtion.
}
\examples{
addData2Nodes(bam.test.obj, cluster.info)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
