% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{bam.process}
\alias{bam.process}
\title{Bam File Process Function}
\usage{
bam.process(bam.file, pattern, short.nt.before.tag, short.nt.after.tag)
}
\arguments{
\item{bam.file}{The input bam data directory}

\item{pattern}{The pattern to seek for}

\item{short.nt.before.tag}{A short sequence before the 8nt tag to help more specific identification}

\item{short.nt.after.tag}{A short sequence after the 8nt tag to help more specific identification}
}
\value{
A data table contains cell barcode, celltag and umi information
}
\description{
This function extracts CellTags from the bam sequencing file, provides cell barcode, umi and their corresponding celltag information.
}
\examples{
bam.process("data.fastq", "CCGGT[ATCG]{8}GAATTC", "CCGGT", "GAATTC")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
