% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Barcode.Aggregate}
\alias{Barcode.Aggregate}
\title{CellTag Barcode Aggregation function}
\usage{
Barcode.Aggregate(file.list, output.file)
}
\arguments{
\item{file.list}{files in a list to aggregate in order same as the BAM files}

\item{output.file}{where to save this aggregated output file. Should be a .tsv file.}
}
\value{
A list containing the pattern, nucleotides to look for before/after the motif
}
\description{
This function allows barcode aggregation of multiple-file processing.
}
\examples{
Barcode.Aggregate(list("barcodes_1.tsv", "barcodes_2.tsv"), output.file = "barcode_aggr.tsv")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
