% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagForCollapsing.R
\name{CellTagDataForCollapsing}
\alias{CellTagDataForCollapsing}
\title{CellTag Starcode Prior Collapsing}
\usage{
CellTagDataForCollapsing(celltag.obj, output.file)
}
\arguments{
\item{celltag.obj}{A CellTag object with the raw count matrix filled.}

\item{output.file}{The filepath and name to save the table for collapsing (usually a .txt file)}
}
\value{
A CellTag object with collapsing mapping table stored in pre.starcode slot
}
\description{
This function generate the .txt file that will be fed into starcode - https://github.com/gui11aume/starcode - to collapse similar CellTags.
}
\examples{
CellTagDataForCollapsing(bam.test.obj, "./collapsing.txt")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
