% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagForCollapsing.R
\name{CellTagDataPostCollapsing}
\alias{CellTagDataPostCollapsing}
\title{CellTag Starcode Post Collapsing}
\usage{
CellTagDataPostCollapsing(celltag.obj, collapsed.rslt.file)
}
\arguments{
\item{celltag.obj}{A CellTag object with the pre-starcode mapping matrix filled.}

\item{collapsed.rslt.file}{File path to the collapsed result file}
}
\value{
A CellTag object with collapsed count matrix stored in collapsed.count slot
}
\description{
This function processes the result generated from starcode - https://github.com/gui11aume/starcode.
}
\examples{
CellTagDataPostCollapsing(bam.test.obj, "./collapsing_result.txt")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
