% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagMatrixGeneration.R
\name{CellTagMatrixCount}
\alias{CellTagMatrixCount}
\title{CellTag Matrix Generation Function}
\usage{
CellTagMatrixCount(celltag.obj, barcodes.file)
}
\arguments{
\item{celltag.obj}{A CellTag object with bam file result filled}

\item{barcodes.file}{A .tsv output file from 10x CellRanger pipeline. It contains a list of all cell barcodes identified in the filtered dataset.}
}
\value{
A CellTag object with the attribute (raw.count) filled
}
\description{
This function uses the extract information from data processed before and generate a Cell Barcode x CellTag matrix
}
\examples{
CellTagMatrixCount(bam.test.obj, "barcodes.tsv")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
