% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCellTagObject.R
\name{CellTagObject}
\alias{CellTagObject}
\title{Create a New CellTag Object}
\usage{
CellTagObject(object.name, fastq.bam.directory)
}
\arguments{
\item{object.name}{The name of the object}

\item{fastq.bam.input}{The input fastq/bam data directory}

\item{celltag.version}{Which version of CellTags are you working with?}
}
\value{
A CellTag Object with open attributes that can be filled as analysis moving along
}
\description{
This function creates a CellTag object that contains the basic information required for the object
}
\examples{
CellTagObejct("hf1.d15.test", "hf1.d15.bam", "v1")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
