% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{CellTagPatternCalling}
\alias{CellTagPatternCalling}
\title{CellTag Pattern Calling Function}
\usage{
CellTagPatternCalling(celltag.version)
}
\arguments{
\item{celltag.version}{Which CellTag version are you investigating?}
}
\value{
A list containing the pattern, nucleotides to look for before/after the motif
}
\description{
This function provides motif patterns corresponding to the input celltag version
}
\examples{
CellTagPatternCalling("v1")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
