% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagWhitelistGeneration.R
\name{CellTagWhitelistFiltering}
\alias{CellTagWhitelistFiltering}
\title{CellTag Whitelist Filtering Function}
\usage{
CellTagWhitelistFiltering(celltag.obj, percentile, output.dir = NULL)
}
\arguments{
\item{celltag.obj}{A CellTag Object with CellTag frequency table counted and sorted}

\item{percentile}{A fraction cutoff percentile for filtering the CellTags e.g. 0.9 for 90th percentile}

\item{output.dir}{Which directory would you like to store these files? If NULL, save to the same directory as the fastq/bam file}
}
\value{
A CellTag Object with attribute (whitelist) filled.
}
\description{
This function conducts whitelist filtering such that only CellTags with count number over their certain percentile would be considered for clone calling
}
\examples{
CellTagWhitelistFiltering(bam.test.obj, 0.9)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
