% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneCalling.R
\name{CloneCalling}
\alias{CloneCalling}
\title{Clone Calling Function}
\usage{
CloneCalling(celltag.obj, correlation.cutoff)
}
\arguments{
\item{celltag.obj}{A CellTag object with the jaccard matrix generated}

\item{correlation.cutoff}{Correlation cutoff for clone membership}
}
\value{
A CellTag object with attributes (clone.composition & clone.size.info) filled.
}
\description{
This function conducts clone calling based on the Jaccard results.
}
\examples{
CloneCalling(bam.test.obj, 0.7)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
