% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagNetworkContruction.R
\name{convertCellTagMatrix2LinkList}
\alias{convertCellTagMatrix2LinkList}
\title{Convert CellTag Matrix to Link List}
\usage{
convertCellTagMatrix2LinkList(celltag.obj)
}
\arguments{
\item{celltag.obj}{A CellTag object with all clone information filled}
}
\value{
A CellTag object with the attribute (network.link.list) filled
}
\description{
This function convert the CellTag Matrix to a link list, which is further used for network construction and visualizetion
}
\examples{
convertCellTagMatrix2LinkList(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
