% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagNetworkVisualiztion.R
\name{drawSubnet}
\alias{drawSubnet}
\title{Draw the Network}
\usage{
drawSubnet(celltag.obj, tag, overlay)
}
\arguments{
\item{celltag.obj}{A CellTag object with link list and nodes filled}

\item{tag}{Which tags would you like to plot?}

\item{overlay}{What information would you like to overlay with the network? This should be one of the column names of the node information.}
}
\value{
A CellTag object with the attribute (network) modified.
}
\description{
This function generate a force-directed network based on the link list and nodes information.
}
\examples{
drawSubnet(bam.test.obj, "CellTagV1_2", "Cluster")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
