% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{fastq.process}
\alias{fastq.process}
\title{Fastq Process Function}
\usage{
fastq.process(fastq.file, pattern, short.nt.before.tag, short.nt.after.tag)
}
\arguments{
\item{fastq.file}{The input fastq/bam data directory}

\item{pattern}{The pattern to seek for}

\item{short.nt.before.tag}{A short sequence before the 8nt tag to help more specific identification}

\item{short.nt.after.tag}{A short sequence after the 8nt tag to help more specific identification}
}
\value{
A list contains count table of CellTags. If requested to save fullTag counts, i.e. save.fullTag.counts = TRUE, return a list of both 8nt tags and full sequences count. Otherwise, a list of 8nt tags counts.
}
\description{
This function extracts CellTags from the raw fastq sequencing file, provides counts of each CellTag and sorts them in desending order.
}
\examples{
fastq.process("data.fastq", "CCGGT[ATCG]{8}GAATTC", "CCGGT", "GAATTC")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
