% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagNetworkContruction.R
\name{getNodesfromLinkList}
\alias{getNodesfromLinkList}
\title{Get Nodes from Link List}
\usage{
getNodesfromLinkList(celltag.obj)
}
\arguments{
\item{celltag.obj}{A CellTag object with link list filled}
}
\value{
A CellTag object with the attribute (nodes) filled
}
\description{
This function extracts the node information from the generated link list.
}
\examples{
getNodesfromLinkList(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
