% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneCalling.R
\name{JaccardAnalysis}
\alias{JaccardAnalysis}
\title{Jaccard Analysis Function}
\usage{
JaccardAnalysis(celltag.obj, plot.corr = TRUE)
}
\arguments{
\item{celltag.obj}{A CellTag object with the counts filtered based on metrics}

\item{plot.corr}{Would you like to plot the correlation matrix?}
}
\value{
A CellTag object with attribute (jaccard.mtx) filled
}
\description{
This function conducts Jaccard analysis to calculate the Jaccard similarity between cells.
}
\examples{
JaccardAnalysis(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
