% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricBasedPlottingAndFiltering.R
\name{MetricBasedFiltering}
\alias{MetricBasedFiltering}
\title{Metric-Base Filtering Function}
\usage{
MetricBasedFiltering(celltag.obj, cutoff, comparison = "less")
}
\arguments{
\item{celltag.obj}{A CellTag Object with count matrix generated}

\item{cutoff}{The cutoff decided from the metric plots}

\item{comparison}{Would you like to maintain the part less than/greater than the cutoff? Default to less. Choices can be greater or less.}
}
\value{
A CellTag Object with attribute (metric.filtered.count) filled
}
\description{
This function applies further filtering on scRNA-seq data with CellTags based on cutoff values identified from the metric plots.
}
\examples{
MetricBasedFiltering(bam.test.object, 20, "less")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
