% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricBasedPlottingAndFiltering.R
\name{MetricPlots}
\alias{MetricPlots}
\title{CellTag Metric Plotting Function}
\usage{
MetricPlots(celltag.obj)
}
\arguments{
\item{celltag.obj}{A CellTag Object}
}
\description{
This function provides some metric plots for further downstream celltag filtering in the scRNA-seq dataset.
}
\examples{
MetricPlots(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
