% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagExtraction.R
\name{CellTagExtraction}
\alias{CellTagExtraction}
\title{CellTag Extraction Function}
\usage{
CellTagExtraction(celltag.obj, celltag.version)
}
\arguments{
\item{celltag.obj}{A CellTag object initialized with path to the fastq/bam file}
}
\value{
A CellTag object with attribute (bam.parse.rslt) filled
}
\description{
This function extracts CellTags from the raw fastq/bam sequencing file. If it is a fastq file, provides counts of each CellTag and sorts them in desending order. If it is a bam file, returns the barcode, umi, celltag information.
}
\examples{
CellTagExtraction(bam.test.obj)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
